Numbers drawn to spots
I had big problem with this where if drawnNumbers is not an array, it would draw an index of that array, what I did is made them vectors and put vector.size() in for loop instead of the actual size number. Now it seems to work but I have a problem, in some cases one number is missing.