Added that you can gameAcces with enter (not muvh)
Will add password hiding with replacing a last character with a dot (password dot), but there is a problem that it actually changes the string which would be like "...ds23" and that is not what we typed. Will work this out in futre. GOTTA DO THE TIME FOR DRAWING BALLS!