V0.1
Overall it's not that bad, it is just basic, I am just a beginner in SFML so I don't really know how to use it's full power, so I am not gonna judge it on library, it was fun, stressful but enjoying journey. If I was doing it every day for couple of hours with a great plan, I would have done the full version so far. I am satisfied for now. Will work on it more soon.

Got to add tickets (60% done) and results at the end (40% done)